DEBUG_PLOT(): use println() to finish line DEBUG_
The current way to finish a line makes the header be finished:

  FIELDA,FIELDB,...\n\r

and the values:

  VALA1,VALB1,...\n
  VALA2,VALB2,...\n
  VALA3,VALB3,...\n

which confuses tools that need to tell if the format of the text file
is DOS/Windows (\r\n) or Unix (\n).

This patch makes it be consistent by using println() to finish lines
in both cases.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>