sensors: add patient state modeling
To improve measurements and more accurately align with the patient,
model the patient behavior in a state machine similar to the control
state machine of breathing in and out. Adjust all the measured values
to use this new state machine instead of the control state machine
for determining when to record.

Signed-off-by: Rosen, Michael R <michael.r.rosen@intel.com>